Make table editable; accept parameters
