+++ src/pages/index.tsx
@@ -2,6 +2,8 @@ import useJaneHopkins from '../hooks/useJaneHopkins';
 import ResponsiveAppBar from '@/components/appbar';
 import { Typography } from '@mui/material';
 import DataTable from '@/components/data-table';
+import { GridColDef, GridColumns, GridRowModel, GridRowsProp } from '@mui/x-data-grid';
+import React from 'react';
 
 export default function Home() {
   const { entities } = useJaneHopkins();
@@ -15,6 +17,21 @@ export default function Home() {
     });
     console.log(addPatientResponse);
   }
+
+  //////////////
+  // Example local data
+  const columns: GridColDef[] = [
+    { field: 'id', headerName: 'ID', width: 10 },
+    { field: 'name', headerName: 'Name', width: 250 },
+    { field: 'age', headerName: 'Age', width: 10 }
+  ]
+  const rows = [
+    { id: 1, name: 'James A', age: 1 },
+    { id: 2, name: 'James B', age: 2 },
+    { id: 3, name: 'James asdasdasdB', age: 3 },
+  ]
+  //////////////
+
   return (
     <>
       <ResponsiveAppBar />
@@ -26,7 +43,10 @@ export default function Home() {
       >Add Patient
       </button>
       <Typography variant='h3' align='center'>Example table</Typography>
-      <DataTable />
+      <DataTable
+        columns={columns}
+        rows={rows}
+      />
     </>
   );
 }
